+++ changelog.md
@@ -4,3 +4,4 @@
   * Addition of Czech translation
   * Addition of Portuguese translation
   * Addition of an editing area before the scan button allowing to immediately read the text that has just been scanned
+  * Add a parameter to disable / enable automatic reading of a document, go to the NVDA settings menu then Scanvox for NVDA